Add layers to the net
Start building up the needed layers for the network to actually contain
anything. At the moment it doesn't do anything. Nothing runs. But it's a
start to fleshing out the final path. You basically only see the input
and output layers. You should not be able to mess with the hidden layers
once you have your network setup.

Might create a "LayerManager" or something that actually creates the
layers that are needed and routes one layers output to the next layers
output. This would also allow all the layers to be package private
instead of public. Just expose the manager instead. Might be worth
doing.

Also unsure how much state the Neurons should hold on to. I mean, they
will not be visible by anyone. And having them keep state might make
sense since we need to store weights somewhere. And the neuron is the
logical place to store it. But I don't like creating states... They are
the reason we need reboots when things mess up.