Add LayerManager and simple app usage
Basically just proves that it does *something*. It's not really learning
anything. But it works, which is nice.

Work to be done is figure out how I want to update weights and bias. The
neurons are statefull. Even though I might not like state, it's what
they have to be. The neuron has the state and that is what it has
'learned'. That state is what makes ANNs powerful. So just deal with
it.

Anyhow. Added a "LayerManager" because I don't want to expose the
layers. The LayerManager is exposed though.